backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
distro==1.9.0
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
exceptiongroup==1.2.2
geographiclib==2.0
geopy==2.4.1
googlemaps==4.10.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
openai==1.59.5
psycopg==3.2.3
psycopg-binary==3.2.3
pydantic==2.10.4
pydantic_core==2.27.2
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0